FIX ElasticaService::define() now allows index to be deleted before (re)defining it again (#20)
* FIX ElasticaService::define() now allows index to be deleted before (re)defining it again

This fixes a bug where changing mapping types on a DataObject after the index
has been created would result in a fatal error, but allowing you to explicitly
delete the index before redefining it again.

* Add Travis build configuration, phpcs ruleset and phpunit test suite configuration

* FIX Remove PHP 7 specific code syntax to restore PHP 5.6 compatibility

PHP 5.6 is listed as the minimum requirement in composer.json, this restores that.
Fixes https://github.com/heyday/silverstripe-elastica/issues/15

* API Remove PHP 7 return type hint

This is a breaking change, but the method is proteted. Up to the maintainers as
to whether they are happy to merge this. We can also bump the minimum version
of PHP to ^7.0 instead.