Use ShowInSearch rather than getField()
getField(), unlike hasField(), does not check for the corresponding
getter method. This is because getField is itself meant to be called
from getter methods. In other words, hasField and getField are not
exactly parallel, by design.

This causes any DataObject whose ShowInSearch exists only as a getter to
be unconditionally deleted from the index whenever it is saved.

This fix works by replacing the getField call with a reference to
ShowInSearch, which will automatically call getShowInSearch.